requirements.txt
python-dotenv
requests
pandas
bs4
matplotlib
plotly
supabase
APScheduler
line-bot-sdk
loguru
